Improve Build & Packaging Setup
Improved the build and packaging setup by migrating to the sole use of
`pyproject.toml` rather than a combination of `pyproject.toml`,
`setup.cfg`, `setup.py` and `MANIFEST.in`, simplifying and
consolidating the build and setup process.

Rearranged the `requirements.*.txt` files, separating the dependencies
into those needed by the library (`requests`), those needed for
development (`purest` and `black`) and those needed for building and
distribution (`build`, `twine` and `wheel`).

Incremented the library version to `1.1.6`.